- CustomMapCollector Initial Implementation \n-gitignore modification
